+++ evm_loader/program/src/instruction/mod.rs
@@ -72,7 +72,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE?]` Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little endian
@@ -87,7 +86,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE?]` Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little endian
@@ -101,7 +99,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE]`  Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little endian
@@ -117,7 +114,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE]`  Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little endian
@@ -132,7 +128,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE]`  Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little endian
@@ -183,7 +178,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE?]` Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little-endian
@@ -198,7 +192,6 @@ pub enum EvmInstruction {
     ///  `[WRITE]` Treasury
     ///  `[WRITE]` Operator Balance
     ///  `[]` System program
-    ///  `[]` Sysvar account
     ///  `[WRITE?]` Other accounts
     /// Instruction data:
     ///  0..4 - treasury index in little-endian
@@ -274,6 +267,7 @@ pub mod config_get_property_count;
 pub mod config_get_status;
 pub mod config_get_version;
 pub mod create_main_treasury;
+pub mod dynamic_fee_transaction_validator;
 pub mod neon_tokens_deposit;
 pub mod operator_create_balance;
 pub mod operator_delete_balance;
