+++ evm_loader/program/src/instruction/operator_withdraw_balance.rs
@@ -21,13 +21,6 @@ pub fn process<'a>(
     operator_balance.validate_owner(&operator)?;
     operator_balance.withdraw(&mut target_balance)?;
 
-    let accounts_db = AccountsDB::new(
-        &[],
-        operator,
-        Some(operator_balance),
-        Some(system),
-        None,
-        None,
-    );
+    let accounts_db = AccountsDB::new(&[], operator, Some(operator_balance), Some(system), None);
     target_balance.increment_revision(&Rent::get()?, &accounts_db)
 }
