+++ evm_loader/program/src/instruction/transaction_cancel.rs
@@ -27,14 +27,7 @@ pub fn process<'a>(
     log_data(&[b"HASH", transaction_hash]);
     log_data(&[b"MINER", operator_balance.address().as_bytes()]);
 
-    let accounts_db = AccountsDB::new(
-        &accounts[3..],
-        operator,
-        Some(operator_balance),
-        None,
-        None,
-        None,
-    );
+    let accounts_db = AccountsDB::new(&accounts[3..], operator, Some(operator_balance), None, None);
     let (storage, _) = StateAccount::restore(program_id, &storage_info, &accounts_db)?;
 
     validate(&storage, transaction_hash)?;
