+++ evm_loader/program/src/instruction/transaction_step.rs
@@ -9,6 +9,7 @@ use crate::evm::tracing::NoopEventListener;
 use crate::evm::{ExitStatus, Machine};
 use crate::executor::{Action, ExecutorState, ExecutorStateData};
 use crate::gasometer::{Gasometer, LAMPORTS_PER_SIGNATURE};
+use crate::instruction::dynamic_fee_transaction_validator;
 use crate::types::boxx::boxx;
 use crate::types::TransactionPayload;
 use crate::types::TreeMap;
@@ -194,7 +195,7 @@ fn finalize<'a, 'b>(
     // (2) charge the User in favor of Operator with amount of `priority_fee_per_gas` * `LAMPORTS_PER_SIGNATURE`.
     if let TransactionPayload::DynamicFee(dynamic_fee_payload) = &storage.trx().transaction {
         // Validate.
-        accounts.db().sysvar().validate_priority_fee(
+        dynamic_fee_transaction_validator::validate_priority_fee(
             dynamic_fee_payload.max_priority_fee_per_gas,
             dynamic_fee_payload.max_fee_per_gas,
         )?;
