+++ evm_loader/program/src/instruction/transaction_step_from_instruction.rs
@@ -1,8 +1,7 @@
 use crate::account::legacy::{TAG_HOLDER_DEPRECATED, TAG_STATE_FINALIZED_DEPRECATED};
 use crate::account::{
     program, AccountsDB, AccountsStatus, Holder, Operator, OperatorBalanceAccount,
-    OperatorBalanceValidator, StateAccount, Sysvar, Treasury, TAG_HOLDER, TAG_STATE,
-    TAG_STATE_FINALIZED,
+    OperatorBalanceValidator, StateAccount, Treasury, TAG_HOLDER, TAG_STATE, TAG_STATE_FINALIZED,
 };
 use crate::debug::log_data;
 use crate::error::{Error, Result};
@@ -31,17 +30,15 @@ pub fn process<'a>(
     let treasury = Treasury::from_account(program_id, treasury_index, &accounts[2])?;
     let operator_balance = OperatorBalanceAccount::try_from_account(program_id, &accounts[3])?;
     let system = program::System::from_account(&accounts[4])?;
-    let sysvar = Sysvar::from_account(&accounts[5])?;
 
     operator_balance.validate_owner(&operator)?;
 
     let accounts_db = AccountsDB::new(
-        &accounts[6..],
+        &accounts[5..],
         operator.clone(),
         operator_balance.clone(),
         Some(system),
         Some(treasury),
-        Some(sysvar),
     );
 
     let mut excessive_lamports = 0_u64;
